+++ src/parsers/response/mod.rs
@@ -12,6 +12,7 @@
 pub mod calendar_graph_response;
 pub mod city_response;
 pub mod date_grid_response;
+pub mod deals_response;
 pub mod explore_response;
 pub mod flight_response;
 pub mod offer_response;